Move scan tests with ArrayK
